Update to TFT_eSPI latest version 1.4.21 (2019/11/12). and SMOOTH_FONT extension enabled.

+++ src/Fonts/GFXFF/gfxfont.h
@@ -12,7 +12,7 @@
 #ifdef LOAD_GFXFF
 
 typedef struct { // Data stored PER GLYPH
-	uint16_t bitmapOffset;     // Pointer into GFXfont->bitmap
+	uint32_t bitmapOffset;     // Pointer into GFXfont->bitmap
 	uint8_t  width, height;    // Bitmap dimensions in pixels
 	uint8_t  xAdvance;         // Distance to advance cursor (x axis)
 	int8_t   xOffset, yOffset; // Dist from cursor pos to UL corner
